python/requirements.txt
tencentcloud-cls-sdk-python>=1.0.0
pyyaml>=5.0
